change FilledTriangle drawing to make it more consistent (#159)
* change sort_points to sort_points_horizotal_vertical and use it for Line

Earlier the sorting was done first vertically and if they were the same,
then horizontally. But for drawing a triangle, we will do it the other
way round. Also, doing it first horizontally and then vertically ensures
that the linear index always keeps increasing

* add ConsistentFilledTriangle and change some previous FixedTriangle
methods to work for AbstractTriangle

* remove VerticalBaseFilledTriangle and FilledTriangle

* rename ConsistentFilledTriangle to FilledTriangle

* change sort_triangle_points to sort_points_horizontal_vertical

* remove type signature from sort_points_horizontal_vertical method in
line.jl

* update test for FilledTriangle

* add docstring for FilledTriangle

* restore line.jl and instead add sort_points_horizontal_vertical in
triangle.jl

will update Line drawing in a separate PR